+++ sdk/eventgrid/eventgrid/src/cadl-generated/api/EventGridContext.ts
@@ -1,24 +1,22 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { ClientOptions } from "@azure-rest/core-client";
 import { EventGridContext } from "../rest/index";
-import { AzureKeyCredential } from "@azure/core-auth";
+import { TokenCredential } from "@azure/core-auth";
 import getClient from "../rest/index";
-import { ClientOptions } from "../common/interfaces";
+
+export interface EventGridClientOptions extends ClientOptions {}
 
 export { EventGridContext } from "../rest/index";
 
 /** Azure Messaging EventGrid Client */
 export function createEventGrid(
   endpoint: string,
-  credential: AzureKeyCredential,
-  options: ClientOptions = {}
+  credential: any | TokenCredential,
+  options: EventGridClientOptions = {}
 ): EventGridContext {
   const baseUrl = endpoint;
-  options.credentials = {
-    ...options.credentials,
-    apiKeyHeaderName: "Authorization",
-  };
   const clientContext = getClient(baseUrl, credential, options);
   return clientContext;
 }