+++ sdk/eventgrid/eventgrid/src/cadl-generated/rest/models.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 /** Properties of an event published to an Azure Messaging EventGrid Namespace topic using the CloudEvent 1.0 Schema. */
-export interface CloudEvent extends Record<string, unknown> {
+export interface CloudEvent {
   /** An identifier for the event. The combination of id and source must be unique for each distinct event. */
   id: string;
   /** Identifies the context in which an event happened. The combination of id and source must be unique for each distinct event. */
@@ -10,7 +10,7 @@ export interface CloudEvent extends Record<string, unknown> {
   /** Event data specific to the event type. */
   data?: unknown;
   /** Event data specific to the event type, encoded as a base64 string. */
-  data_base64?: Uint8Array;
+  data_base64?: string;
   /** Type of event related to the originating occurrence. */
   type: string;
   /** The time (in UTC) the event was generated, in RFC3339 format. */
@@ -25,20 +25,26 @@ export interface CloudEvent extends Record<string, unknown> {
   subject?: string;
 }
 
-/** Array of lock token strings for the corresponding received Cloud Events to be acknowledged. */
+/** Array of lock tokens for the corresponding received Cloud Events to be acknowledged. */
 export interface AcknowledgeOptions {
-  /** String array of lock tokens. */
+  /** Array of lock tokens. */
   lockTokens: string[];
 }
 
-/** Array of lock token strings for the corresponding received Cloud Events to be released. */
+/** Array of lock tokens for the corresponding received Cloud Events to be released. */
 export interface ReleaseOptions {
-  /** String array of lock tokens. */
+  /** Array of lock tokens. */
   lockTokens: string[];
 }
 
-/** Array of lock token strings for the corresponding received Cloud Events to be rejected. */
+/** Array of lock tokens for the corresponding received Cloud Events to be rejected. */
 export interface RejectOptions {
-  /** String array of lock tokens. */
+  /** Array of lock tokens. */
+  lockTokens: string[];
+}
+
+/** Array of lock tokens for the corresponding received Cloud Events to be renewed. */
+export interface RenewLockOptions {
+  /** Array of lock tokens. */
   lockTokens: string[];
 }